requirements.txt
annotated-types==0.5.0
anyio==3.7.1
cffi==1.15.1
click==8.1.6
cryptography==41.0.2
exceptiongroup==1.1.2
fastapi==0.100.0
h11==0.14.0
idna==3.4
numpy==1.25.1
oracledb==1.3.2
pandas==2.0.3
pycparser==2.21
pydantic==2.0.3
pydantic-settings==2.0.2
pydantic_core==2.3.0
pypyodbc==1.3.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
tzdata==2023.3
uvicorn==0.23.1
